moved components to a separate files
