+++ src/components/TodoList/TodoList.tsx
@@ -1,4 +1,30 @@
-// Add the required types and props
-export const TodoList = () => (
-  <>TodoList markup</>
-);
+import { Todo } from '../../types/Todo';
+import { TodoInfo } from '../TodoInfo';
+
+interface TodosList {
+  todosList: Todo[];
+}
+
+export const TodoList: React.FC<TodosList> = ({ todosList }) => {
+  return (
+    <section className="TodoList">
+      {todosList.map(todoItem => <TodoInfo todoInfo={todoItem} />)}
+
+      <article className="TodoInfo TodoInfo--completed">
+        <h2 className="TodoInfo__title">JS</h2>
+
+        <a className="UserInfo" href="mailto:Shanna@melissa.tv">
+          Ervin Howell
+        </a>
+      </article>
+
+      <article className="TodoInfo">
+        <h2 className="TodoInfo__title">React</h2>
+
+        <a className="UserInfo" href="mailto:Nathan@yesenia.net">
+          Clementine Bauch
+        </a>
+      </article>
+    </section>
+  );
+};
