+++ src/components/UserInfo/UserInfo.tsx
@@ -1,4 +1,18 @@
-// Add the required types and props
-export const UserInfo = () => (
-  <>UserInfo markup</>
-);
+import { User } from '../../types/User';
+
+interface TodoUser {
+  user: User;
+}
+
+export const UserInfo: React.FC<TodoUser> = ({ user }) => {
+  const {
+    name,
+    email,
+  } = user;
+
+  return (
+    <a className="UserInfo" href={email}>
+      {name}
+    </a>
+  );
+};
